// shared/src/peer/channel-type.enum.ts
export enum ChannelType {

  /**
   * ECG data channel.
   */
  ecg = 'ecg',

  /**
   * EDA data channel.
   */
  eda = 'eda',

  /**
   * EyeTracker data and stream channel.
   */
  eyetracking = 'etk',

  /**
   * Monitoring data channel.
   */
  monitoring = 'mon',

  /**
   * Movement sensor data channel.
   */
  movement = 'mov'
}